+++ pkg/config/sources.go
@@ -1,0 +1,15 @@
+// Copyright (c) 2026 Benjamin Borbe All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
+package config
+
+// FieldSources records which config layer provided each of the 4 layered user-pref fields.
+// Valid values for each field are: "default", "global", "project", "arg".
+// Zero value (empty string) is treated the same as "default" by callers.
+type FieldSources struct {
+	HideGit            string
+	AutoRelease        string
+	DirtyFileThreshold string
+	Model              string
+}